bug(test-coverage): fix dropped coverage
- fixed test coverage dropping after doing some refactoring

[Fixes #163675429]
