+++ src/components/route/index.js
@@ -8,7 +8,8 @@ import App from '../App';
 import Navbar from '../presentation/Navbar';
 import Login from '../auth/login/Login';
 import Signup from '../auth/signup/Signup';
-import { getCurrentUser, getCartItemsCount } from '../../actions';
+import { getCurrentUser } from '../../actions/auth';
+import { getCartItemsCount } from '../../actions/cart';
 import Order from '../order/Order';
 import OrderHistory from '../order/OrderHistory';
 import AddEditMenu from '../menu/AddEditMenu';
